requirements.txt
django_haystack==3.2.1
pandas==1.4.2
PyPDF2==2.11.0
requests==2.27.1
streamlit==1.13.0
~treamlit==1.12.2
